put isolated dir into github_env
