+++ integration/presets.ts
@@ -1,4 +1,3 @@
-/* eslint-disable ts/no-namespace */
 /* eslint-disable node/prefer-global/process */
 import { applicationConfig } from './models/application-config'
 import { fixtures } from './fixtures'
@@ -6,15 +5,6 @@ import { fixtures } from './fixtures'
 const constants = {
   INTEGRATION_PM_NAME: process.env.INTEGRATION_PM_NAME,
   INTEGRATION_PM_VERSION: process.env.INTEGRATION_PM_VERSION,
-}
-
-declare global {
-  namespace NodeJS {
-    interface ProcessEnv {
-      INTEGRATION_PM_NAME?: 'npm' | 'pnpm' | 'yarn' | 'bun'
-      INTEGRATION_PM_VERSION?: string
-    }
-  }
 }
 
 const kitchenSink = applicationConfig()
